% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSheet.R
\name{DataSheet}
\alias{DataSheet}
\title{This is the DataSheet class}
\description{
DataSheet class description
}
\examples{
#DataSheet$new(name="Bill", hair="Blond")
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data_changed}}{Title}

\item{\code{metadata_changed}}{Title}

\item{\code{variables_metadata_changed}}{Title}

\item{\code{current_filter}}{Title}

\item{\code{current_column_selection}}{Title}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSheet-new}{\code{DataSheet$new()}}
\item \href{#method-DataSheet-clone}{\code{DataSheet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-new}{}}}
\subsection{Method \code{new()}}{
Initalise function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$new(
  data = data.frame(),
  data_name = "",
  variables_metadata = data.frame(),
  metadata = list(),
  imported_from = "",
  messages = TRUE,
  convert = TRUE,
  create = TRUE,
  start_point = 1,
  filters = list(),
  column_selections = list(),
  objects = list(),
  calculations = list(),
  keys = list(),
  comments = list(),
  keep_attributes = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{data_name}}{data name}

\item{\code{variables_metadata}}{TODO}

\item{\code{metadata}}{TODO}

\item{\code{imported_from}}{TODO}

\item{\code{messages}}{TODO}

\item{\code{convert}}{TODO}

\item{\code{create}}{TODO}

\item{\code{start_point}}{TODO}

\item{\code{filters}}{TODO}

\item{\code{column_selections}}{TODO}

\item{\code{objects}}{TODO}

\item{\code{calculations}}{TODO}

\item{\code{keys}}{TODO}

\item{\code{comments}}{TODO}

\item{\code{keep_attributes}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
initialise TODO
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
